Split tests into individual files for easier navigation
