--- test/mint.ts
@@ -1,440 +1,0 @@
-import { expect } from 'chai';
-import { ethers } from 'hardhat';
-import hre from 'hardhat';
-import Web3 from 'web3';
-import { deployments, getNamedAccounts } from 'hardhat';
-import { sign } from 'crypto';
-
-import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
-import {
-  CxipRegistry,
-  CxipAssetProxy,
-  CxipCopyrightProxy,
-  CxipERC721Proxy,
-  CxipERC1155Proxy,
-  CxipIdentityProxy,
-  CxipProvenanceProxy,
-  PA1DProxy,
-  CxipProvenance,
-  CxipIdentity,
-  CxipERC721,
-  CxipAsset,
-  PA1D,
-  CxipFactory,
-} from '../typechain';
-
-const web3 = new Web3(Web3.givenProvider || 'ws://localhost:8545');
-
-const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000';
-const error = function (err: string) {
-  console.log(err);
-  process.exit();
-};
-
-const remove0x = function (input: string) {
-  if (input.startsWith('0x')) {
-    return input.substring(2);
-  } else {
-    return input;
-  }
-};
-
-// Format collection name and symbol as bytes32 left padded zero bytes
-// Example: 0x0000000000000000000000000000000000436f6c6c656374696f6e206e616d65
-// NOTE: Similar functionality can be accomplised with web3.utils.fromAscii or ethers.utils.toUtf8Bytes + hexlify
-// But this handles prepending 0x and correct padding with zeros
-// See: https://ethereum.stackexchange.com/questions/96884/string-to-hex-in-ethers-js
-function utf8ToBytes32(str: string) {
-  return (
-    '0x' +
-    Array.from(str)
-      .map((c) =>
-        c.charCodeAt(0) < 128
-          ? c.charCodeAt(0).toString(16)
-          : encodeURIComponent(c).replace(/\%/g, '').toLowerCase()
-      )
-      .join('')
-      .padStart(64, '0')
-  );
-}
-
-describe('CXIP', () => {
-  let deployer: SignerWithAddress;
-  let user: SignerWithAddress;
-  let user2: SignerWithAddress;
-  let user3: SignerWithAddress;
-  let user4: SignerWithAddress;
-
-  let registry: CxipRegistry;
-
-  let assetProxy: CxipAssetProxy;
-  let copyrightProxy: CxipCopyrightProxy;
-  let erc721Proxy: CxipERC721Proxy;
-  let erc1155Proxy: CxipERC1155Proxy;
-  let identityProxy: CxipIdentityProxy;
-  let provenanceProxy: CxipProvenanceProxy;
-  let royaltiesProxy: PA1DProxy;
-
-  let asset: CxipAsset;
-  let erc721: CxipERC721;
-  let identity: CxipIdentity;
-  let provenance: CxipProvenance;
-  let royalties: PA1D;
-  let factory: CxipFactory;
-
-  before(async () => {
-    const accounts = await ethers.getSigners();
-    deployer = accounts[0];
-    user = accounts[1];
-    user2 = accounts[2];
-    user3 = accounts[3];
-    user4 = accounts[4];
-    const {
-      CxipRegistry,
-      CxipAssetProxy,
-      CxipCopyrightProxy,
-      CxipERC721Proxy,
-      CxipERC1155Proxy,
-      CxipIdentityProxy,
-      CxipProvenanceProxy,
-      PA1DProxy,
-      CxipProvenance,
-      CxipIdentity,
-      CxipERC721,
-      CxipERC1155,
-      CxipCopyright,
-      CxipAsset,
-      PA1D,
-      Register,
-    } = await deployments.fixture([
-      'CxipRegistry',
-      'CxipAssetProxy',
-      'CxipCopyrightProxy',
-      'CxipERC721Proxy',
-      'CxipERC1155Proxy',
-      'CxipIdentityProxy',
-      'CxipProvenanceProxy',
-      'PA1DProxy',
-      'CxipProvenance',
-      'CxipIdentity',
-      'CxipERC721',
-      'CxipERC1155',
-      'CxipCopyright',
-      'CxipAsset',
-      'PA1D',
-
-      'Register',
-    ]);
-    registry = await ethers.getContract('CxipRegistry');
-    assetProxy = await ethers.getContract('CxipAssetProxy');
-    copyrightProxy = await ethers.getContract('CxipCopyrightProxy');
-    erc721Proxy = await ethers.getContract('CxipERC721Proxy');
-    erc1155Proxy = await ethers.getContract('CxipERC1155Proxy');
-    identityProxy = await ethers.getContract('CxipIdentityProxy');
-    provenanceProxy = await ethers.getContract('CxipProvenanceProxy');
-    royaltiesProxy = await ethers.getContract('PA1DProxy');
-
-    provenance = await ethers.getContract('CxipProvenance');
-    identity = await ethers.getContract('CxipIdentity');
-    erc721 = await ethers.getContract('CxipERC721');
-    asset = await ethers.getContract('CxipAsset');
-    royalties = await ethers.getContract('PA1D');
-  });
-
-  beforeEach(async () => {});
-
-  afterEach(async () => {});
-
-  describe('Registry', () => {
-    it('should set and get asset source', async () => {
-      const assetTx = await registry.setAssetSource(asset.address).catch(error);
-      await assetTx.wait();
-      const assetSourceAddress = await registry.getAssetSource();
-
-      expect(assetSourceAddress).to.equal(asset.address);
-    });
-
-    it('should set and get asset proxy', async () => {
-      const assetProxyTx = await registry
-        .setAsset(assetProxy.address)
-        .catch(error);
-      await assetProxyTx.wait();
-      const assetProxyAddress = await registry.getAsset();
-      expect(assetProxyAddress).to.equal(assetProxy.address);
-    });
-
-    it('should set and get ERC721', async () => {
-      const erc721Tx = await registry
-        .setERC721CollectionSource(erc721.address)
-        .catch(error);
-      await erc721Tx.wait();
-      const erc721Address = await registry.getERC721CollectionSource();
-      expect(erc721Address).to.equal(erc721.address);
-    });
-
-    it('should set and get identity', async () => {
-      const identityTx = await registry
-        .setIdentitySource(identity.address)
-        .catch(error);
-      await identityTx.wait();
-      const identitySourceAddress = await registry.getIdentitySource();
-      expect(identitySourceAddress).to.equal(identity.address);
-    });
-
-    it('should set and get provenance', async () => {
-      const provenanceTx = await registry
-        .setProvenanceSource(provenance.address)
-        .catch(error);
-      await provenanceTx.wait();
-      const provenanceAddress = await registry.getProvenanceSource();
-      expect(provenanceAddress).to.equal(provenance.address);
-    });
-
-    it('should set and get provenance proxy', async () => {
-      const provenanceProxyTx = await registry
-        .setProvenance(provenanceProxy.address)
-        .catch(error);
-      await provenanceProxyTx.wait();
-      const provenanceProxyAddress = await registry.getProvenance();
-      expect(provenanceProxyAddress).to.equal(provenanceProxy.address);
-    });
-
-    it('should set and get royalties', async () => {
-      const royaltiesTx = await registry
-        .setPA1DSource(royalties.address)
-        .catch(error);
-      await royaltiesTx.wait();
-      const royaltiesAddress = await registry.getPA1DSource();
-      expect(royaltiesAddress).to.equal(royalties.address);
-    });
-
-    it('should set and get royalties proxy', async () => {
-      const royaltiesProxyTx = await registry
-        .setPA1D(royaltiesProxy.address)
-        .catch(error);
-      await royaltiesProxyTx.wait();
-      const royaltiesProxyAddress = await registry.getPA1D();
-      expect(royaltiesProxyAddress).to.equal(royaltiesProxy.address);
-    });
-  });
-
-  describe('Identity', () => {
-    it('should create an identity', async () => {
-      const salt = deployer.address + '0x000000000000000000000000'.substring(2);
-      const tx = await provenance.createIdentity(
-        salt,
-        '0x' + '00'.repeat(20), // zero address
-        [
-          '0x0000000000000000000000000000000000000000000000000000000000000000',
-          '0x0000000000000000000000000000000000000000000000000000000000000000',
-          '0x0',
-        ] as any
-      );
-
-      const receipt = await tx.wait();
-      const events = receipt.events?.filter((x) => {
-        return x.event == 'IdentityCreated';
-      });
-
-      const identityAddress = await provenance.getIdentity();
-      expect(identityAddress).not.to.equal(ZERO_ADDRESS);
-      // expect(identityAddress).to.equal(events?[0].args[0]);
-      // expect(identityAddress).to.equal(events?[0].args.identityAddress);
-    });
-
-    it('should not allow duplicate identities', async () => {
-      const salt = deployer.address + '0x000000000000000000000000'.substring(2);
-
-      await expect(
-        provenance.createIdentity(
-          salt,
-          '0x' + '00'.repeat(20), // zero address
-          [
-            '0x0000000000000000000000000000000000000000000000000000000000000000',
-            '0x0000000000000000000000000000000000000000000000000000000000000000',
-            '0x0',
-          ] as any
-        )
-      ).to.be.revertedWith('CXIP: wallet already used');
-    });
-
-    it.skip('should create an identity with a secondary wallet', async () => {
-      // TODO: Come back to this and use signatures for second address
-    });
-
-    it('should be a valid identity', async () => {
-      const salt = user.address + '0x000000000000000000000000'.substring(2);
-      const tx = await await provenance.connect(user).createIdentity(
-        salt,
-        '0x' + '00'.repeat(20), // zero address
-        [
-          `0x000000000000000000000000${user.address.substring(2)}`,
-          `0x000000000000000000000000${user.address.substring(2)}`,
-          '0x0',
-        ] as any
-      );
-
-      const receipt = await tx.wait();
-      const identityAddress = await provenance.getIdentity();
-      expect(identityAddress).not.to.equal(ZERO_ADDRESS);
-
-      const isValid = await provenance.isIdentityValid(identityAddress);
-      expect(isValid).to.equal(true);
-    });
-
-    it('should not be a blacklisted identity', async () => {
-      const salt = user2.address + '0x000000000000000000000000'.substring(2);
-      const tx = await await provenance.connect(user2).createIdentity(
-        salt,
-        '0x' + '00'.repeat(20), // zero address
-        [
-          `0x000000000000000000000000${user2.address.substring(2)}`,
-          `0x000000000000000000000000${user2.address.substring(2)}`,
-          '0x0',
-        ] as any
-      );
-
-      const receipt = await tx.wait();
-      const identityAddress = await provenance.getIdentity();
-      expect(identityAddress).not.to.equal(ZERO_ADDRESS);
-
-      const isValid = await provenance.isIdentityBlacklisted(identityAddress);
-      expect(isValid).to.equal(false);
-    });
-  });
-
-  describe('Collection', () => {
-    it('should create a collection', async () => {
-      const salt = user3.address + '0x000000000000000000000000'.substring(2);
-
-      // Attach the provenance implementation ABI to provenance proxy
-      const p = await provenance.attach(provenanceProxy.address);
-      const tx = await p.connect(user3).createIdentity(
-        salt,
-        '0x' + '00'.repeat(20), // zero address
-        [
-          `0x000000000000000000000000${user3.address.substring(2)}`,
-          `0x000000000000000000000000${user3.address.substring(2)}`,
-          '0x0',
-        ] as any
-      );
-
-      const receipt = await tx.wait();
-      const identityAddress = await p.connect(user3).getIdentity();
-
-      // Attach the identity implementation ABI to the newly created identity proxy
-      const i = await identity.attach(identityAddress);
-
-      const result = await i.connect(user3).createERC721Collection(
-        salt,
-        user3.address,
-        [
-          `0x0000000000000000000000000000000000000000000000000000000000000000`,
-          `0x0000000000000000000000000000000000000000000000000000000000000000`,
-          '0x0',
-        ] as any,
-        [
-          `${utf8ToBytes32('Collection name')}`, // Collection name
-          '0x0000000000000000000000000000000000000000000000000000000000000000', // Collection name 2
-          `${utf8ToBytes32('Collection symbol')}`, // Collection symbol
-          user3.address, // royalties (address)
-          '0x0000000000000000000003e8', // 1000 bps (uint96)
-        ] as any
-      );
-
-      result.wait();
-
-      const collectionAddress = await i.getCollectionById(0);
-      const collectionType = await i.getCollectionType(collectionAddress);
-      expect(collectionAddress).not.to.equal(ZERO_ADDRESS);
-      expect(collectionType).not.to.equal(ZERO_ADDRESS);
-    });
-
-    it.skip('should create a collection using a different wallet in the identity', async () => {});
-  });
-
-  describe('ERC721', () => {
-    it('should create a ERC721 NFT in a collection', async () => {
-      const salt = user4.address + '0x000000000000000000000000'.substring(2);
-
-      // Attach the provenance implementation ABI to provenance proxy
-      const p = await provenance.attach(provenanceProxy.address);
-      const tx = await p.connect(user4).createIdentity(
-        salt,
-        '0x' + '00'.repeat(20), // zero address
-        [
-          `0x000000000000000000000000${user4.address.substring(2)}`,
-          `0x000000000000000000000000${user4.address.substring(2)}`,
-          '0x0',
-        ] as any
-      );
-
-      const receipt = await tx.wait();
-      const identityAddress = await p.connect(user4).getIdentity();
-
-      // Attach the identity implementation ABI to the newly created identity proxy
-      const i = await identity.attach(identityAddress);
-      const result = await i.connect(user4).createERC721Collection(
-        salt,
-        user4.address,
-        [
-          `0x0000000000000000000000000000000000000000000000000000000000000000`,
-          `0x0000000000000000000000000000000000000000000000000000000000000000`,
-          '0x0',
-        ] as any,
-        [
-          `${utf8ToBytes32('Collection name')}`, // Collection name
-          '0x0000000000000000000000000000000000000000000000000000000000000000', // Collection name 2
-          `${utf8ToBytes32('Collection symbol')}`, // Collection symbol
-          user4.address, // royalties (address)
-          '0x0000000000000000000003e8', // 1000 bps (uint96)
-        ] as any
-      );
-
-      result.wait();
-
-      const collectionAddress = await i.getCollectionById(0);
-      const collectionType = await i.getCollectionType(collectionAddress);
-      expect(collectionAddress).not.to.equal(ZERO_ADDRESS);
-      expect(collectionType).not.to.equal(ZERO_ADDRESS);
-
-      //////////////////////////////////////////////////////////////////////////////
-      // Create an NFT
-      //////////////////////////////////////////////////////////////////////////////
-      /**
-       * This signature composition is required to send in the payload to create ERC721
-       */
-      const payload =
-        '0x398d6a45a2c3d1145dfc3a229313e4c3b65165eb0b8b04c0fe787d0e32924775';
-      const tokenId =
-        '0x0000000000000000000000000000000000000000000000000000000000000001';
-      const wallet = user4.address;
-      const sig = await user4.signMessage(payload);
-      const signature = {
-        r: '0x' + sig.substring(2, 66),
-        s: '0x' + sig.substring(66, 130),
-        v: '0x' + (parseInt('0x' + sig.substring(130, 132)) + 27).toString(16),
-      };
-
-      const arHash = 'd3dStWPKvAsticf1YqNT3FQCzT2nYlAw' + 'RVNFVlKmonc';
-      const ipfsHash = 'QmX3UFC6GeqnmBbthWQhxRW6WgTmWWVd' + 'ist3TL59UbTZYx';
-      const nftTx = await i
-        .connect(user4)
-        .createERC721Token(collectionAddress, tokenId, [
-          payload,
-          [signature.r, signature.s, signature.v] as any,
-          wallet,
-          web3.utils.asciiToHex(arHash.substring(0, 32)),
-          web3.utils.asciiToHex(arHash.substring(32, 43)),
-          web3.utils.asciiToHex(ipfsHash.substring(0, 32)),
-          web3.utils.asciiToHex(ipfsHash.substring(32, 46)),
-        ] as any);
-
-      try {
-        await nftTx.wait();
-      } catch (error: any) {
-        throw new Error(error);
-      }
-    });
-  });
-});
